Register board to resource directory on startup
When the board starts it tries to register to a resource directory
until the registering succeeds.
After that an update request is sent to the resource directory
periodically (period defined in Makefile).
When the update fails only a message is printed so far because
trying to register again like on boot seems not to work (I have to
figure out what the issue is).